+++ STACpopulator/implementations/CMIP6_UofT/add_CMIP6.py
@@ -1,29 +1,20 @@
 import argparse
 import json
-import logging
 from datetime import datetime
-from typing import Any, List, Literal, MutableMapping, Optional
+from typing import Any, List, Literal, MutableMapping, NoReturn, Optional
 
 import pydantic_core
 import pyessv
-from colorlog import ColoredFormatter
+from requests.sessions import Session
 from pydantic import AnyHttpUrl, ConfigDict, Field, FieldValidationInfo, field_validator
 from pystac.extensions.datacube import DatacubeExtension
 
+from STACpopulator.cli import add_request_options, apply_request_options
 from STACpopulator.implementations.CMIP6_UofT.extensions import DataCubeHelper
 from STACpopulator.input import GenericLoader, ErrorLoader, THREDDSLoader
 from STACpopulator.models import GeoJSONPolygon, STACItemProperties
 from STACpopulator.populator_base import STACpopulatorBase
-from STACpopulator.stac_utils import STAC_item_from_metadata, collection2literal
-
-LOGGER = logging.getLogger(__name__)
-LOGFORMAT = "  %(log_color)s%(levelname)s:%(reset)s %(blue)s[%(name)-30s]%(reset)s %(message)s"
-formatter = ColoredFormatter(LOGFORMAT)
-stream = logging.StreamHandler()
-stream.setFormatter(formatter)
-LOGGER.addHandler(stream)
-LOGGER.setLevel(logging.INFO)
-LOGGER.propagate = False
+from STACpopulator.stac_utils import LOGGER, STAC_item_from_metadata, collection2literal
 
 # CMIP6 controlled vocabulary (CV)
 CV = pyessv.WCRP.CMIP6
@@ -108,7 +99,13 @@ class CMIP6populator(STACpopulatorBase):
     item_properties_model = CMIP6ItemProperties
     item_geometry_model = GeoJSONPolygon
 
-    def __init__(self, stac_host: str, data_loader: GenericLoader, update: Optional[bool] = False) -> None:
+    def __init__(
+        self,
+        stac_host: str,
+        data_loader: GenericLoader,
+        update: Optional[bool] = False,
+        session: Optional[Session] = None,
+    ) -> None:
         """Constructor
 
         :param stac_host: URL to the STAC API
@@ -116,7 +113,7 @@ class CMIP6populator(STACpopulatorBase):
         :param thredds_catalog_url: the URL to the THREDDS catalog to ingest
         :type thredds_catalog_url: str
         """
-        super().__init__(stac_host, data_loader, update)
+        super().__init__(stac_host, data_loader, update=update, session=session)
 
     @staticmethod
     def make_cmip6_item_id(attrs: MutableMapping[str, Any]) -> str:
@@ -159,9 +156,9 @@ class CMIP6populator(STACpopulatorBase):
 
         # Add datacube extension
         try:
-            dchelper = DataCubeHelper(item_data)
+            dc_helper = DataCubeHelper(item_data)
             dc_ext = DatacubeExtension.ext(item, add_if_missing=True)
-            dc_ext.apply(dimensions=dchelper.dimensions, variables=dchelper.variables)
+            dc_ext.apply(dimensions=dc_helper.dimensions, variables=dc_helper.variables)
         except Exception:
             LOGGER.warning(f"Failed to add Datacube extension to item {item_name}")
 
@@ -169,23 +166,37 @@ class CMIP6populator(STACpopulatorBase):
         return json.loads(json.dumps(item.to_dict()))
 
 
-if __name__ == "__main__":
-    parser = argparse.ArgumentParser(prog="CMIP6 STAC populator")
+def make_parser() -> argparse.ArgumentParser:
+    parser = argparse.ArgumentParser(description="CMIP6 STAC populator")
     parser.add_argument("stac_host", type=str, help="STAC API address")
     parser.add_argument("thredds_catalog_URL", type=str, help="URL to the CMIP6 THREDDS catalog")
     parser.add_argument("--update", action="store_true", help="Update collection and its items")
+    parser.add_argument("--mode", choices=["full", "single"],
+                        help="Operation mode, processing the full dataset or only the single reference.")
+    add_request_options(parser)
+    return parser
 
-    args = parser.parse_args()
 
-    LOGGER.info(f"Arguments to call: {args}")
+def runner(ns: argparse.Namespace) -> Optional[int] | NoReturn:
+    LOGGER.info(f"Arguments to call: {vars(ns)}")
 
-    mode = "full"
+    with Session() as session:
+        apply_request_options(session, ns)
+        if ns.mode == "full":
+            data_loader = THREDDSLoader(ns.thredds_catalog_URL, session=session)
+        else:
+            # To be implemented
+            data_loader = ErrorLoader()
 
-    if mode == "full":
-        data_loader = THREDDSLoader(args.thredds_catalog_URL)
-    else:
-        # To be implemented
-        data_loader = ErrorLoader(args.error_file)
+        c = CMIP6populator(ns.stac_host, data_loader, update=ns.update, session=session)
+        c.ingest()
 
-    c = CMIP6populator(args.stac_host, data_loader, args.update)
-    c.ingest()
+
+def main(*args: str) -> Optional[int]:
+    parser = make_parser()
+    ns = parser.parse_args(args or None)
+    return runner(ns)
+
+
+if __name__ == "__main__":
+    main()